Add Dia diagram and Ingredient and Pizza classes.
Added a Dia diagram which contains the database model.
The Ingredient class provides a base class for all ingredients which
will go into the pizza. There are four child classes: Crust, Sauce,
Cheese, and Topping. The child classes currently do not have their own
properties, but are subclassed to in case we need to add properties
later.